+++ nerdlets/slo-r-main/components/index.js
@@ -1,2 +1,2 @@
-export { default as Overview } from './overview/overview-view';
+export { default as SloCombine } from './slo-combine/main-view';
 export { default as SloList } from './slo-list/main-view';
